pyproject.toml
[project]
name = "agentic-crew"
version = "0.2.0"
description = "Framework-agnostic AI crew orchestration - declare once, run on CrewAI, LangGraph, or Strands"
requires-python = ">=3.11"
license = { text = "MIT" }
readme = "README.md"
authors = [{ name = "Jon Bogaty", email = "jon@jonbogaty.com" }]
keywords = ["crewai", "langgraph", "strands", "ai-agents", "multi-agent", "orchestration"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
# Core dependencies - NO framework required (auto-detected at runtime)
dependencies = [
    "pyyaml>=6.0.3",
    "pydantic>=2.12.4",
]

[project.optional-dependencies]
# === AI Framework Extras (install one or more) ===
# CrewAI - full featured, default choice
crewai = [
    "crewai[tools,anthropic]>=1.5.0",
]
# LangGraph - graph-based flows
langgraph = [
    "langgraph>=0.2.0",
    "langchain-anthropic>=0.3.0",
]
# AWS Strands - lightweight, AWS-native
strands = [
    "strands-agents>=0.1.0",
]
# All frameworks
ai = [
    "crewai[tools,anthropic]>=1.5.0",
    "langgraph>=0.2.0",
    "langchain-anthropic>=0.3.0",
    "strands-agents>=0.1.0",
]

# === Feature Extras ===
# MCP server support
mcp = [
    "mcp>=1.0.0",
]
# Web scraping tools (for connector builder)
scraping = [
    "crewai[tools]>=1.5.0",
    "beautifulsoup4>=4.12.0",
]
# Meshy integration
meshy = [
    "vendor-connectors[meshy]>=0.2.0",
]

# === Development ===
tests = [
    "pytest>=9.0.1",
    "pytest-mock>=3.15.1",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-timeout>=2.4.0",
]
dev = [
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]
docs = [
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=2.0.0",
    "sphinx-autodoc-typehints>=2.0.0",
    "myst-parser>=3.0.0",
]

# === Full Installation ===
all = [
    # All frameworks
    "crewai[tools,anthropic]>=1.5.0",
    "langgraph>=0.2.0",
    "langchain-anthropic>=0.3.0",
    "strands-agents>=0.1.0",
    # Features
    "mcp>=1.0.0",
    "beautifulsoup4>=4.12.0",
    "vendor-connectors[meshy]>=0.2.0",
]

[project.scripts]
agentic-crew = "agentic_crew.main:main"
run-crew = "agentic_crew.crew:kickoff"
train-crew = "agentic_crew.crew:train"

[project.urls]
Homepage = "https://github.com/jbcom/agentic-crew"
Repository = "https://github.com/jbcom/agentic-crew"
Documentation = "https://github.com/jbcom/agentic-crew#readme"
Issues = "https://github.com/jbcom/agentic-crew/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agentic_crew"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "C4"]
ignore = ["D", "T201"]

[tool.ruff.lint.isort]
known-first-party = ["agentic_crew"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
